backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
starlette==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
cryptography==41.0.8

# AI and ML
google-generativeai==0.3.2

# File processing
pypdf==3.17.4
python-docx==1.1.0
openpyxl==3.1.2
Pillow==10.1.0
python-magic==0.4.27
aiofiles==23.2.0

# HTTP and networking
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1
websockets==12.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7
rich==13.7.0

# Caching and performance
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0

# Monitoring and logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Email
fastapi-mail==1.4.1
jinja2==3.1.2

# CORS and middleware
fastapi-cors==0.0.6
slowapi==0.1.9

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
